requirements.txt
streamlit
python-dotenv
autogen
groq